backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=4.0.1",
    "email-validator>=2.0.0",
    "fastapi>=0.123.5",
    "httpx>=0.28.1",
    "jwt>=1.4.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.7",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
